all: rename internal package to pkg
Pretty much all of pkgsite is inside of an "internal" package, which
prevents any other code from using it. Additionally, some of those
subpackages then have their own "internal" packages. This renames the
top-level "internal" package to "pkg" so that it can be used outside of
this repo.

This is the key change that makes this repo "pkgsitelib" rather than
just "pkgsite".